server: improve cluster controller correctness
We noticed that informers based on CRDs need to be started *after* the
CRDs exist and are established. Also, looping over _all_ contexts in the
server-generated kubeconfig is no longer correct, and we should choose
the specific ones we needed. Of note, looping over all was creating CRDs
with cluster=*.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>