Make OpenAPIV3Schema atomic
Set the x-kubernetes-map-type to atomic for the OpenAPIV3Schema field in
apiresourceimports and negotiatedapiresources. This causes server-side
apply to treat the schema as an atomic unit, instead of trying to track
each individual field/element of the schema (which causes 34-second
timeouts in the apiserver trying to process the data).

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>