test: e2e: an informer-driven expectation helper
In tests, we want to focus on describing the state we expect the system
to fall into, not worrying about handling watch events and the like.
This commit adds a simple informer-driven expectation helper that allows
callers to provide some expectation and have it evaluated when the state
of the world changes.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>