workspaceshard: record the version of loaded credentials
Downstream consumers of shard credentials need to know when the
credentials are rotated. For instance, any client that first determines
that data they care about is on a shard, then begins some process to
monitor that data, must know when the credentials used to access that
shard change. In order to not require that all of these clients watch
Secrets, record the version of the credentials that are loaded.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>