crdpuller: use a RESTMapper
Use a RESTMapper to select the prioritized resource by name in the event
of multiple partial matches. For example, given
--resources-to-sync=pods, and if you have the metrics-server and its
APIs installed, this will match pods (core) and pods (metrics.k8s.io).
Using the RESTMapper allows us to select the resource with the highest
priority.

NOTE: this implementation is not efficient and is just a quick
demonstration!

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>