authz: use workspaceLister from Server
Use the workspaceLister from the Server startup because we only enable
it if the workspace controllers are running. That's the only way we
currently have to guarantee the Workspace CRD is installed. Otherwise,
our post start hook that starts the shared informers will hang waiting
for caches to sync (if the Workspace CRD isn't installed).

The side effect is that authz will not check for Workspace existence if
the workspace scheduler/controllers aren't running. This is a temporary
fix, most likely.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>