syncer: fix hotloop & minor cleanup
Fix a bug in the syncer where, after it got an error, it failed to
return immediately after adding the key back to the queue with a rate
limit. Because it didn't return immediately, the key was subsequently
forgotten (i.e. rate limit reset), which broke rate limited backoff.

Refactor the holder struct not to store an object instance, which is
consistent with standard controller/queue best practices.

Update the process method code to (hopefully) correctly determine the
"from" and "to" namespaces and item keys.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>