e2e: Ensure artifacts are saved at test completion for shared fixture
Previously, the new sharable fixture was running artifact-collecting
functions registered at setup but not those registered during test
execution. When executed in parallel, a sub-test's modification of a
shared fixture's set of artifact functions wasn't propagating to the
parent test.

This change switches to registering artifact functions with
t.Cleanup() to ensure that artifact collection is tied to the scope of
the registering test rather than the scope of the fixture lifetime. It
also ensures that artifacts for subtests using shared fixture are
saved in a path specific to the subtest.