Refactor namespace scheduler cluster assignment for testability
This change extracts cluster selection from the namespace scheduling
controller and adds preliminary unit test coverage of the result.

API interaction is limited to functions that can be replaced in
tests. Care was taken to ensure functional behavior which is easier to
reason about and validate. Testing of API mutation is out-of-scope
since the fixture cost would likely exceed the benefit.

This strategy of factoring out logic that can be unit tested in this
fashion enables trivial validation of corner-cases and is intended to
reduce the cost of ongoing maintenance.