Support adding labels to workspace on creation
A field `AdditionalWorkspaceLabels` is added to
`ClusterWorkspaceType`. Labels defined by this field for a given
`ClusterWorkspaceType` will be added to `ClusterWorkspace` resources
of that type. Labels will only be added on creation of a
`ClusterWorkspace` and only if the labels are not already
defined (i.e. existing values will not be overwritten).