kubectl-kcp: point to authInfo config instead of clone
Currently when running e.g. kubectl kcp workspaces use my-workspace, the
context created in the user's kubeconfig will point to a new user that
was cloned from the previously-active context. This can fill a user's
kubeconfig with lots of duplicate users if there are lots of workspace
switches. It also allows the authInfo to drift from the original context
and the workspace-specific context, whereas they very likely need to be
in sync.

This change will make the new context point to the already-existing
authInfo instead of making a copy of it.

Note that the kubeconfig authInfo will be updated for overrides.

Co-authored-by: Chris Sams csams@redhat.com

Signed-off-by: Kyle Lape klape@redhat.com