crd bootstrap: better goroutine support
Add support for multiple goroutines (e.g. post start hooks) that try to
install the same CRD at the same time. There was a race where one
goroutine checked for the CRD's existence, didn't find it, and proceeded
to create it. Meanwhile, another goroutine created it. Before this
change, the first goroutine would return an error ("already exists").
Now, it recognizes the existence and tries to do an update instead of
returning an error.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>