Make namespace scheduling opt-in via label on containing workspace
Not all use of kcp will involve namespace scheduling or
synchronization with a workload cluster.

One such use case involves using a logical cluster (workspace) as a
workload cluster for the purposes of syncer testing. Were namespace
scheduling to remain enabled for such a workspace, the cluster label
of namespaces created by the syncer in that workspace would be
removed (due to the cluster not existing in the workspace). Since the
syncer's downstream informer filters on the presence of the label in
question, the syncer would not be able to see any resources for which
the namespace scheduler had removed the label. Disabling the namespace
scheduler for the workspace ensures the contents of the workspace more
precisely reflect how a kubernetes cluster would behave.

Unlike disabling scheduling for a namespace - where status is still
set by the scheduler on the namespace - the goal here is disabling the
scheduler entirely. To that end, only log messages will indicate that
scheduling is disabled for the workspace.