Create new external.kubeconfig
external.kubeconfig uses the --external-hostname option to populate the
hostname *and port*.  The apiserver config allows the port to be added to the
--external-hostname.

admin.kubeconfig will now always use the IP address of the locally bound
interface and the port from --secure-port.

--kubeconfig-path now points to the directory containing the kubeconfig
files.  This now means that the admin.kubeconfig filename cannot be
changed.

Fixes #707

Signed-off-by: Kyle Lape <klape@redhat.com>