Ensure deployment env vars are mutated correctly by push-mode syncer
The push mode syncer configures itself to talk to kcp over localhost,
and the deployment mutator was using that same localhost configuration
to set env vars allowing a deployment to talk back to kcp. If the
cluster that a deployment was synced to was on a different host,
though, that deployment wouldn't be able to talk to kcp as intended.

This change ensures the push mode syncer is explicitly provided the
external address that synced deployments should be updated with rather
than extracting it from the kubeconfig the syncer uses to talk to kcp.