Ensure the kcp config root has reasonable permissions.
Some of the components that kcp starts write private configuration into
the configuration root before the kcp server comes up. This means that
to create the configuration directory with the permissions we want,
we need to do it really early so we aren't affected by these side-effects.

For the embedded etcd specifically, we should use the etcd "TouchDirAll"
API that is specifically provided for creating the configuration root.
It's also useful to make the etcd state persist in a subdirectory to
avoid any possibility of file name collisions.

Signed-off-by: James Peach <jpeach@cloudflare.com>