Add front proxy
This adds the front proxy into the topology of the deployment manifest.

Instead of terminating TLS at the ingress router, it will instead
terminate at the front proxy, and the proxy will route requests to
either KCP or the virtual workspaces process.

This also leverages cert-manager to add the ability to authenticate using
a certificate.  Groups will be assigned using the organizations listed
on the cert.