Namespace scheduler improvements
When reconciling a resource, if its namespace can't be found in the
cache, return nil instead of an error (presumably the delete event for
the resource is coming shortly).

When enqueuing resources for a namespace, because namespace-scoped
listers currently return all items for namespace X across all
logical clusters, filter out any item that is for a different logical
cluster.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>