virtual: snip dependency on CRD storage builder
We don't need every possible endpoint to implement rest.StandardStorage
in every case, so we can refactor to reflect that. It seems like we're
getting a couple things from the current custom resource REST:

- shallow-copies of metadata, needed in the past when the generic
  registry was mutating it to set self-links, no longer needed
- providing categories
- forcing no creation on update behavior

By not using the custom resource REST implementation, we need to
duplicate ~10 lines of code, which does not seem risky.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>