pkg/authorization: don't check for cluster workspace lister.
Currently checking for initialize permissions on the parent cluster relies nil-checking the cluster workspace lister.
This worked under the assumption that clusterworkspace CRDs are being installed in the workspace controller/scheduler.
This does not hold true any more. Clusterworkspace CRDs are now initialized as part of the server bootstrap,
independent of the workspace controller/scheduler, hence these CRDs are always available.
Even if we omit bootstrapping those CRDs the value will not be nil (verified in a debugger session).

This fixes it.