Fix partial meta request determination logic
The accept header can have multiple clauses, but we were only supporting
one. The Kubernetes metadataSharedInformerFactory generates requests
with the accept header having multiple clauses, and our previous logic
didn't handle them correctly, leading to thinking the request was not
for partial metadata, when in fact it was supposed to be.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>