apibindings: fix partial metadata bug
Fixing isPartialMetadataRequest led to the discovery of a bug in the
apiBindingAwareCRDLister's Get() method. The apibinding deletor uses the
partial metadata client to list resources to check for remaining data
(i.e. CRs with finalizers). Without the fix to
isPartialMetadataRequest, the metadata client was sending an Accept
header that led us to say it was *not* a partial metadata request. As
such, the apiBindingAwareCRDLister would proceed to getForFullData when
looking up a CRD, meaning it would find a CRD via an APIBinding, which
is what we want in this case.

With the fix, however, now the deletor's metadata client requests are
correctly interpreted as partial metadata requests. The
apiBindingAwareCRDLister's logic was finding the first CRD matching the
group and resource, which is less and less likely to find the correct
CRD as the number of workspaces, bindings, and CRDs grows.

The solution is to update the special handling of partial metadata
requests when trying to get the correct CRD. We want to handle
the following cases (in order of priority):

1. System CRD
2. CRD via identity (which could be for a specific cluster, or wildcard
- doesn't matter)
3. Wildcard
  a. For partial metadata
  b. Non partial metadata
4. In a specific cluster - check APIBindings first, then local CRDs

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>