e2e: require two failed pings to fail the test
To whatever extent our test setup is causing etcd to have a hiccup here
or there, it does not seem to then cause the rest of the tests to have
any issues. While it's useful to know that there are some latent etcd
issues (or whatever) in the tests, we should be more tolerant of probing
failures, since things happen and distributed, eventually consistent
systems are hard.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>