ddsif: prepare for quota support
Add DiscoveryData() to the ddsif. This will be used by Kubernetes quota.

Add Subscribe/Unsubscribe methods to notify subscribers when
informers/discovery change. This will be used by quota.

Rename InformerForResource to ForResource to make it compatible with
informerfactory.InformerFactory.

For non-deleted CRDs, only notify that we need to check for informers to
add/remove if the CRD in the watch event is Established or if its
Established condition changed.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>