e2e/clusterworkspacedeletion: deflake
Wait for the default namespace to exist before trying to delete it. The
default namespace is created async when a workspace is created, and in
some cases, the e2e test code that is trying to delete the namespace
executes before the namespace has been created, resulting in an
unexpected not-found error.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>