introduce NewWildcardIdentitiesWrappingRoundTripper
NewWildcardIdentitiesWrappingRoundTripper creates a HTTP RoundTripper that injected resource identities for
individual group or group resources. Each group or resource is coming from one APIExport whose
names are passed in as a map.
The RoundTripper is exposed as a function that allows wrapping the RoundTripper.

The method also returns the resolve function that gets the APIExports and extract the identities.
The resolve func might return an error if the APIExport is not found or for other reason. Only
after it succeeds a client using the returned RoundTripper can use the group and group resources
with identities.