change the storage prefix to /cache
change the storage prefix under which all resources are kept
this allows us to store the same GR under a different
prefix than the kcp server (/registry).
It is useful when this server shares the database with a kcp instance.

It boils down to the following on the storage level:
for listing across shard:  /cache/<group>/<resource>:<identity>/*
for listing for one shard: /cache/<group>/<resource>:<identity>/<shard>/*