Generally, the RawPath is only available when some chars have been encoded. The Path always fully expands the URL.
Since we trim the Path, naturally we would like also to trim the RawPath. To do that correctly we would have to consider encoded slash char as well (%2F).

For example the code breaks on %2Fclusters%2F*/apis/apis.kcp.dev/v1alpha1/apiexports

Because we don't really need to deal with encoded chars and we don't use the RawPath only the Path we could simplify.
This PR simply recomputes the URL after trimming and overwrites it in the HTTP requests.