pkg/virtual/workspaces/registry: fix panic sending to a closed channel
Currently, the projection watch interface closes its result channel in Stop().
This can cause a panic when the goroutine started in ResultChan() still tries
to send data to the result channel which can be closed.

This fixes it by making sure that the goroutine started in ResultChan()
itself closes the channel once the delegate watch is done.