syncer: bifurcate the namespace controller
Today, one controller/workqueue is used to drive two separate
reconciliation flows: reacting to events in the upstream and downstream
informers. The presence of a cluster in the queue key is used to infer
which informers the event came from.

This is hard to reason about, and the logic for the two cases is
entirely separate, and, furthermore, when we're syncing to a kcp logical
cluster, the "downstream" objects have a cluster set and the keys do not
follow the expected format.

Instead of inferring things from stringified keys in one queue, we can
just split the two controllers and explicitly use cluster-aware or
cluster-unaware keyfunctions where necessary.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>