ci: don't install a local registry with kind
For test-e2e-shared and test-e2e-sharded, we need a kind cluster and
for our images (syncer, test-image) to be available from that
cluster.

In github, we use the container-tools/kind-action to set up this kind
cluster, which also sets up a local registry by default and configures
the kind cluster to use it.

It turns out, if you want to reproduce this in your dev environment,
you don't need to have this local registry setup because ko will
happily side-load the image into the kind node. This seems pretty
convenient.

Since the local registry isn't necessary, and probably not what you
want when developing, let's not use it in CI for consistency.

See:
https://github.com/container-tools/kind-action/blob/main/registry.sh
https://github.com/ko-build/ko/blob/main/pkg/publish/kind/write.go