Stop shortening git sha used to update RH gitops repo
This is to support the use of Openshift CI images, which are tagged
using the full git sha instead of the shortened one.

Note that the deploy job will still wait for the image to be built in
Github Actions before applying the change to the gitops repo.  This can
be addressed in a follow up PR to remove the build job when we are
ready.