✨ pkg/virtual/apiexport: add maximal permission policy authorization (#1882)
* pkg/virtual/apiexport: add maximal permission policy authorization

Currently, maximal permission policy is not enforced in the virtual apiexport apiserver.
This fixes it.

* docs/authorization: add maximum permission policy

* test/e2e/virtual/apiexport: add e2e test for shadowed CRD

* pkg/server,pkg/virtual/framework/forwardingregistry: allow non-wildcard identity request

Currently, non-wildcard requests having an identity (i.e. virtual apiexport apiserver requests)
are not handled correctly. Formally, a client can execute a non-wildcard request for getting
CRDs in a workspace via an apiexport apiserver virtual workspace.

This fixes it.

* pkg/authorization,pkg/virtual/apiexport/authorizer: add audit logging and anonymization

* pkg/virtual/apiexport/authorizer: add unit tests

* test/e2e/fixtures/apifixtures: eventually create API binding

... because permissions might not have effect yet.

* SQUASH: address review comments