APIResourceImport: deprecate the Location field
This field is used by the syncer and various controllers to identify
which SyncTarget owns the import. However, we can do this using the
SyncTarget owner reference.

Removing this field would decouple the apiresource API group from the
workload API group - i.e. this is currently the only place in the
apiresource types where we refer to a workload concept.

Removing it would also remove a point of SyncTarget vs Location vs
Cluster confusion.