cache: replication controller: attach the shard annotaiton during object creation
to avoid an additional UPDATE request, mismatch on the Generation, CreationTime or UID field, replicated objects will have those fields set.
by setting the shard name during object creation we instruct the cache server to stop updating/changing those fields in any way.
in addition to that, the cache server will remove the annotation before persisting the object
because a shard name is calculated based on the storage prefix.