Revert additional proxy auth methods in e2e-sharded
This is a partial revert of #2178 - which passed e2e-sharded,
but there are reports of that job now taking twice as long
with these new auth methods enabled.

I'm reverting the Makefile changes to restore the CI job to
the previous behavior, and will investigate locally why
performance is impacted.