quota: use a workqueue to manage updating monitors
When there's a lot of churn in the types contained in `kcp`, the
type-aware shared informer factory will be sending out a lot of
notifications in a short window of time. The previous implementation
used a single goroutine to consume the notifications, therefore
restricting the allowed production throughput to the consumption
throughput, leading to notifications being lost.

In order to be more resilient to this type of situation, we need to
decouple consumption from production, and while we're doing that we do
not want to hold on to a large number of pending notifications, nor do
we want to duplicate work when many notifications are received during
the handling of the first.

As the monitor update is idempotent, but not reentrant, we can utilize a
workqueue with one worker to ensure that handling on the consumption is
single-threaded, while de-duplicating notifications that occur while
monitor updates are in flight.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>