Increase CRD cleanup requeue delay; remove e2e test
Requeue is increased from 10 seconds to 30 minutes.

When KCP is under heavy load during e2e test runs, it can take a long
time for a new `APIBinding` to complete reconciliation.  In some cases, it
takes longer than the requeue delay configured in the CRD cleanup
controller, which means the newly-created CRD will be deleted before
initialization of its corresponding `APIBinding` is complete.

The obvious fix is to increase the requeue delay, but there is an e2e
test that waits on this delay to ensure CRDs are cleaned up properly.
Unfortunately waiting becomes infeasible if the delay is increased to the
order of minutes or hours instead of seconds.

Therefore the e2e test is removed in favor of an expanded unit test.
The queue is mocked out in the unit test, and the test function emulates
time going forward by decreasing the CRD's `creationTimestamp` before
calling `process` again.

Fixes #2507