⚠️ Move Workspace.Status.{Cluster, URL} to Spec
In order to preserve the URL field, given that Status may be wiped under
a number of conditions and most times URL cannot be easily or at all
reconstructured, this changeset moves the field under Spec.

In addition, now the admission code also checks that any modification to
the URL is performed by a system user.

Signed-off-by: Vince Prignano <vince@prigna.com>