test/e2e: remove a data race
The value of `ws` previously would have been mutated by the
`Eventually()` call after this cleanup, leading to a race when cleanup
occurred during workspace initialization. We don't expect the name of
the workspace to ever change, so it's appropriate to capture the name
before registering eleanup and simply assuming the name is static.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>