Fix gap in APIExport virtual workspace queues
Fix a gap in the APIExport virtual workspace queueing where it was
possible to incorrectly return a "not found" error when trying to list a
claimed resource. The gap was that we weren't queueing the dependent
APIExports; i.e., if APIExport A gets updated, and APIExport B has a
permission claim on A's identity, we need to make sure to queue B in
this situation.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>