Fix git in Prow
Newer versions of git have a security measure in place to prevent git
from parsing the .git/config file or running any hooks if the user
running git is not the owner of the repository files. Prow runs as a
random UID, which is not guaranteed to be the same as the UID that owns
the cloned files, so this change disables the ownership check and
considers all cloned repositories "safe" regardless of owner.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>