workload resource controller: use actual GVR
The dynamic discovery shared informer factory does a wildcard partial
metadata list/watch against specific API versions. When an informer
event comes in for a specific item, it's for a specific API version
(whatever the factory used), but that doesn't necessarily mean it's the
API version of the actual object being retrieved (the API version
doesn't matter because we're only getting partial metadata).

But if you want to UPDATE/DELETE the object, you must know its appropriate API
version. This is now presented via annotations in kcp.io/original-api-version.
This updates the reconciler to use the annotation when making dynamic client
requests.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>