Fix KCPRequestInfoResolver for kind: Cluster
The cache server preprocesses the path to remove the cluster prefix
in https://github.com/kcp-dev/kcp/blob/88d6071b2aaf47bcd5e854f8209e4c3bd8035b35/pkg/cache/server/config.go#L172-L186.

When a resource of type "cluster" is created and the first cluster
prefix is dropped, the path would be a k8s request info path (/apis/...)
but will still have the `clusters` keyword in the path.

Due to this, the regex in the kcp request info resolver does not work
well.

This commit updates the resolver:

1. If a path is a k8s request info path (begins with `/api` or `/apis`),
then the path is not checked against the regex.

2. The regex is updated to preserve the forward slash in the
strippedURL.