Fix probable bug: imported package name as name identifier
Should be avoided because the package exported identifiers can't be used after the declaration.
This may be also confusing for the reader.

Variable renaming is aligned to the official golang guidelines: 'Variable names in Go should be short rather than long. This is especially true for local variables with limited scope. Prefer c to lineCount. Prefer i to sliceIndex'.
Source: https://go.dev/wiki/CodeReviewComments#variable-names

Signed-off-by: Atanas Alexandrov <sirakov@gmail.com>