feat: add defaultAPIBinding lifecycle controller
Added defaultAPIBinding lifecycle controller to maintain and update
APIBindings declared in WorkspaceTypes.

A user can declare a defaultAPIBindingLifecycle mode as Maintain or
InitializeOnly (default) on their Workspace.
This mode is synced to the LogicalCluster as annotation.

From there, a controller maintains the APIBindings on the LogicalCluster
in sync with the WorkspaceType (and dependencies thereof).
Any newly added defaultAPIBindings are added as APIBinding and newly
declared permission claims are accepted automatically.

Changes by the users (e.g. deleting a maintained APIBinding) are
reconciled by recreating the binding.

on-behalf-of: @eon-se opensource@eon.com
Signed-off-by: Hannes Blut <hannes.blut.external@eon.com>