add manual conversion logic between v1alpha1 and v1alpha2
The conversion ensures that virtual ResourceSchemas are retained in v1alpha1 by using an annotion.
I made the assumption that the order of resource schemas is not important, as during a roundtrip
conversion, non-cRD resource schemas would always be added *after* CRD resource schemas.

On-behalf-of: @SAP christoph.mewes@sap.com