CachedResource replication reconciler should retry on AlreadyExists
If the CachedObject is missing, and Create fails with AlreadyExists,
the reconciler now returns an error and lets the work item to be requeued.
The previous code would, in that case, attempt to do an Update but without
setting the object's meta, the Update call would always fail (missing
resourceVersion and others...). Letting the item to be requeued is cleaner
and fixes object updates.

On-behalf-of: @SAP robert.vasek@sap.com
Signed-off-by: Robert Vasek <robert.vasek@clyso.com>