remove Path option for WorkspaceAuthConfig references
When resolving the authenticators for a type for a workspace for a logicalcluster, it would be very
difficult to resolve authenticators because a workspace doesn't know its own fully qualified path
(a logicalcluster has a special annotation for it, but Workspace objects don't know their parent
path). This then makes it difficult to figure out the default value for authConfig paths, so for now
and for simplicity, we just do not allow AuthConfigs to be in different logicalclusters than the
types referring to them.

This could be solved by using admission and defaulting the path, like kcp does for the type reference
on a Workspace object.

On-behalf-of: @SAP christoph.mewes@sap.com