fix flakiness in TestAPIExportBindingAuthorizer
The test's intention is to ensure there is 1 consumer per shard for the wildwest API. To achieve this
it creates 1 workspace for each shard and then binds to the APIExport from within there. However the
code failed to ensure that the logicalclusters for these workspaces are actually at least evenly
distributed to every shard.

This lead to flakiness as sometimes both logicalclusters were scheduled onto the same shard, and
suddenly there was only one consumer shard and so the test faile.

On-behalf-of: @SAP christoph.mewes@sap.com